L4/main/requirements/base.txt
Django==5.0.3
djangorestframework==3.15.1
drf-yasg==1.21.7
Scrapy==2.11.1
